requirements.txt
# Runtime dependencies
pandas
matplotlib
python-dotenv
google-generativeai
mysql-connector-python

# Test dependencies
pytest
